pyproject.toml
[tool.poetry]
name = "ava"
version = "0.1.0"
description = ""
authors = ["Karan Singh Kochar <pyrotank41@gmail.com>"]
license = "GNU AFFERO GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
openai = "^1.35.7"
loguru = "^0.7.2"
llama-index = "^0.10.51"
llama-index-llms-azure-openai = "^0.1.8"
llama-index-llms-anthropic = "^0.1.15"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
llama-index-vector-stores-qdrant = "^0.2.10"
llama-index-readers-file = "^0.1.25"
qdrant-client = "^1.9.2"
llama-index-readers-google = "^0.2.9"
llama-index-embeddings-azure-openai = "^0.1.10"



[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
